requirements.txt
aiogram
asyncio
httpx
g4f
cachetools
python-dotenv
pytest
pytest-asyncio
allure-pytest
Faker
jsonschema
beautifulsoup4
googletrans-py39
pytest-asyncio
pymorphy2
pymorphy2-dicts
sqlalchemy
asyncpg
psycopg
psycopg_binary
greenlet
aiosqlite
paramiko
jinja2
aiohttp
beautifulsoup4==4.11.1
colorama==0.4.4
Flask
requests==2.28.1
gunicorn
duckduckgo_search
